// src/transactions/transaction-category.enum.ts
export enum TransactionCategory {
  PENDING = 'Pending',
  GROCERIES = 'Groceries',
  DINING_OUT = 'Dining Out',
  UTILITIES = 'Utilities',
  TRANSPORTATION = 'Transportation',
  ENTERTAINMENT = 'Entertainment',
  HEALTHCARE = 'Healthcare',
  SHOPPING = 'Shopping',
  HOUSING = 'Housing',
  EDUCATION = 'Education',
  MISCELLANEOUS = 'Miscellaneous',
}